# openapi.yaml
openapi: 3.1.0
info:
  title: Black Mesa API
  version: '1.0.0'
  description: |
    REST API for the Black Mesa Discord moderation bot.

    All endpoints under `/api/` require a Bearer JWT obtained via the
    `/api/oauth/discord` flow. The token encodes the caller's Discord user
    ID and OAuth tokens and is validated on every request.

    **Permission model**: each protected endpoint checks a specific
    `PermissionSet` flag against the caller.  Flags are a `u64` bitfield —
    see the [Permissions docs](https://docs.blackmesa.bot/bot/permissions/)
    for a full flag reference.
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `/api/oauth/discord`.  Pass as `Authorization: Bearer <token>`.
  schemas:
    Id:
      type: string
      description: Discord snowflake ID (64-bit integer encoded as a decimal string).
      pattern: '^\d+$'
      example: '123456789012345678'
    Uuid:
      type: string
      format: uuid
      example: c488c9ec-423a-437c-817c-41d0d4141e58
...
    ApiError:
      description: Error payload returned by error handlers.
      oneOf:
        - type: string
        - type: object
          additionalProperties: true
...
    InfractionType:
      type: string
      enum:
        - warn
        - mute
        - kick
        - ban
    AutomodOffense:
      type: object
      description: Populated when an infraction was created by the automod engine.
      properties:
        type:
          type: string
          description: One of `message-spam`, `newline-spam`, `word-censor`, `link-censor`, `invite-censor`.
        message:
          type: string
          description: The message content that triggered automod.
        count:
          type: [integer, 'null']
          format: int64
          description: Number of events within the window (spam types only).
        interval:
          type: [integer, 'null']
          format: int64
          description: Window size in milliseconds (spam types only).
        offending_filter:
          type: [string, 'null']
          description: The specific filter string or domain that matched (censor types only).
      example:
        type: message-spam
        message: 'hey hey hey hey hey hey'
        count: 6
        interval: 5000
        offending_filter: null
    Infraction:
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/Uuid'
        guild_id:
          $ref: '#/components/schemas/Id'
        user_id:
          $ref: '#/components/schemas/Id'
        moderator_id:
          $ref: '#/components/schemas/Id'
        infraction_type:
          $ref: '#/components/schemas/InfractionType'
        reason:
          type: [string, 'null']
        last_edited:
          type: [integer, 'null']
          format: int64
          description: Unix timestamp (seconds) of the last edit.
        expires_at:
          type: [integer, 'null']
          format: int64
          description: Unix timestamp (seconds) when this infraction expires and is automatically deactivated.
        mute_role_id:
          oneOf:
            - $ref: '#/components/schemas/Id'
            - type: 'null'
          description: Role ID that was applied for mute infractions.
        automod_offense:
          oneOf:
            - $ref: '#/components/schemas/AutomodOffense'
            - type: 'null'
          description: Set when the infraction was created by the automod engine.
        active:
          type: boolean
          description: Whether the infraction is currently in effect.
      example:
        uuid: c488c9ec-423a-437c-817c-41d0d4141e58
        guild_id: '123456789012345678'
        user_id: '987654321098765432'
        moderator_id: '111222333444555666'
        infraction_type: ban
        reason: Repeatedly violated community rules after multiple warnings.
        active: true
        expires_at: 1775072000
        last_edited: null
        mute_role_id: null
        automod_offense: null
    CreateInfractionRequest:
      type: object
      required:
        - guild_id
        - user_id
        - infraction_type
      properties:
        guild_id:
          $ref: '#/components/schemas/Id'
        user_id:
          $ref: '#/components/schemas/Id'
        infraction_type:
          $ref: '#/components/schemas/InfractionType'
        reason:
          type: [string, 'null']
        expires_at:
          type: [integer, 'null']
          format: int64
          description: Unix timestamp (seconds) when this infraction should expire.
        mute_role_id:
          oneOf:
            - $ref: '#/components/schemas/Id'
            - type: 'null'
          description: Required for `mute` type — the role to apply.
      example:
        guild_id: '123456789012345678'
        user_id: '987654321098765432'
        infraction_type: mute
        reason: Excessive spamming in #general
        expires_at: 1775072000
        mute_role_id: '555666777888999000'
    DeactivateResponse:
      type: object
      required:
        - success
        - id
      properties:
        success:
          type: boolean
        id:
          type: string
          description: UUID of the deactivated infraction.
      example:
        success: true
        id: c488c9ec-423a-437c-817c-41d0d4141e58
paths:
...
  /api/infractions/{guild_id}:
    get:
      summary: List or search infractions for a guild
      description: |
        Returns infractions for a guild. Supports optional filtering by user, type, and
        active status. Requires `INFRACTION_VIEW`.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: guild_id
          required: true
          schema:
            $ref: '#/components/schemas/Id'
        - in: query
          name: user_id
          required: false
          description: Filter to a specific user.
          schema:
            $ref: '#/components/schemas/Id'
        - in: query
          name: type
          required: false
          description: Filter by infraction type.
          schema:
            $ref: '#/components/schemas/InfractionType'
        - in: query
          name: active
          required: false
          description: Filter by active status.
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: Matching infractions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Infraction'
              example:
                - uuid: c488c9ec-423a-437c-817c-41d0d4141e58
                  guild_id: '123456789012345678'
                  user_id: '987654321098765432'
                  moderator_id: '111222333444555666'
                  infraction_type: ban
                  reason: Repeatedly violated community rules.
                  active: true
                  expires_at: 1775072000
                  last_edited: null
                  mute_role_id: null
                  automod_offense: null
                - uuid: d599daf1-534b-548d-928d-52e1e5252f69
                  guild_id: '123456789012345678'
                  user_id: '444555666777888999'
                  moderator_id: null
                  infraction_type: warn
                  reason: Spamming in #general
                  active: true
                  expires_at: null
                  last_edited: null
                  mute_role_id: null
                  automod_offense:
                    type: message-spam
                    message: 'hey hey hey hey hey'
                    count: 6
                    interval: 5000
                    offending_filter: null
        '400':
          description: Invalid query or guild ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized or insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Config not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/infractions:
    post:
      summary: Create a new infraction
      description: |
        Creates an infraction record. The `moderator_id` is automatically set to the
        authenticated user's ID (the request body `moderator_id` field is ignored).
        Requires `INFRACTION_EDIT`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInfractionRequest'
            example:
              guild_id: '123456789012345678'
              user_id: '987654321098765432'
              infraction_type: mute
              reason: Excessive spamming in #general
              expires_at: 1775072000
              mute_role_id: '555666777888999000'
      responses:
        '200':
          description: Created infraction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Infraction'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized or insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Config not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/infractions/{guild_id}/{id}/deactivate:
    post:
      summary: Deactivate an infraction
      description: |
        Marks an infraction as inactive (pardons it).  Returns `400` if the infraction
        is already inactive. Requires `INFRACTION_EDIT`.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: guild_id
          required: true
          schema:
            $ref: '#/components/schemas/Id'
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: Infraction deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeactivateResponse'
              example:
                success: true
                id: c488c9ec-423a-437c-817c-41d0d4141e58
        '400':
          description: Invalid ID or infraction already inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized or insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Config or infraction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'